+++ controllers/users/index.js
@@ -3,7 +3,7 @@ const login = require('./login.js');
 const logout = require('./logout.js');
 const getCurrent = require('./getCurrent.js');
 const updateSubscription = require('./updateSubscription.js');
-const avatar = require('./avatar.js');
+const updateAvatar = require('./updateAvatar.js');
 
 module.exports = {
   register,
@@ -11,5 +11,5 @@ module.exports = {
   logout,
   getCurrent,
   updateSubscription,
-  avatar,
+  updateAvatar,
 };
