+++ package.json
@@ -16,6 +16,7 @@
     "express": "4.17.1",
     "gravatar": "^1.8.2",
     "http-errors": "^2.0.0",
+    "jimp": "^0.22.7",
     "joi": "^17.8.3",
     "jsonwebtoken": "^9.0.0",
     "mongoose": "^7.0.0",
